split tests to dedicated file
